Add saturation display control to user config. Add correct handling of main source of software crash!
miniscopes and cameras in user config now support "showSaturation" bool value to turn on or off red saturation display.

Software now correctly reconnects to a video stream if it temporarily disconnects. This means you can now actually unplug the coax cable and replug it in and the software will continue running.